backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
python-jose[cryptography]
supabase
google-generativeai
pydantic